refactor: move Main in router
