--- src/pages/index.ts
@@ -1,1 +1,0 @@
-export { Main } from './Main';
